refactor(server): extract notify route domain (#594)
Move api_notify into agentwire/routes/notify.py as NotifyRoutesMixin
with register_notify_routes, mirroring the #560 split pattern. Pure
relocation — _post_toast stays on the base class (cross-domain callers).

Closes #594

Built by dotdev.dev
